feature(rejectIfNotFound) added query option
add option to make findById and findOne queries reject their
promise rather than resolve with "null", which could be considered
an illegal/unwanted result.

See https://github.com/Automattic/mongoose/issues/3298, this has
been a long-standing request.  It's my opinion that the increasing
prevalence of promises in lieu of callbacks makes the minor added
overhead of checking this option worthwhile from perspective of
developer efficiency.

The most recent suggested work-around depends on additions to schema
objects, requiring potentially undesirable code repetition.

My new option has parallels to the existing Query.lean option:  the
user wants to tune the semantics of the find operation for a specific
purpose.  For 'lean' this is performance optimization, for 'rejectIfNotFound'
this is for Promise fluency.